remove node_modules from .vscodeignore file
Without this change, no node_modules are included in the vsix (see
the nightly vsix for an example) and the extension does not load.
Based on the number and size of the bundled modules it seems that
the right thing happens anyway and that only needed modules are
bundled.

Note also that this seems to be a new issue, probably due to some
tool upgrades, because I could build and load the extension a few
weeks ago without a problem.